Allow restricting tag generation to files listed by custom commands
This adds a new setting, g:gutentags_file_list_command, which specifies
command(s) to use to list files for which tags should be generated,
instead of recursively examining all files within the project root.

This is useful in projects using source control to restrict tag
generation to only files tracked in the repository.

This setting is conceptually similar to CtrlP's ctrlp_user_command
option.

This implements the feature requested in
https://github.com/ludovicchabant/vim-gutentags/issues/90