Always use the potentially custom name for the tags file.
If the tags file got a custom name from `gutentags_init_user_func`, we need to
use that in all situations. However, the tags file name we get in the
`generate` function is a full path so we make it relative to the project root
if possible when we don't use a cache directory for tags files.