Ads a finder. Adds file list opt. Imoroves doc.
Improved the documentation for gutentags_file_list_command.
Adds the 'types' value to the same dictionary.
Adds the capacity to set g:gutentags_project_info_finder, which
overrides the gutentags_project_info options.
This is the same as the way the g:gutentags_project_root_finder
overrides the gutentags_project_root.