Fix file list command in `update_scope.sh`
Prepend `eval` and surround `$FILE_LIST_CMD` with double quotes to make sure the command actually works if there is single or double quotes in the command. 
For example 'find . -type f -iname "*.php"'. That would fail without the `eval` command. Double quotes are usually needed for when list or pattern is expected, only in a few situation you can omit them safely so I rather put them for safety.