feat: support gutentags_file_list_command on the gtags_cscope module
* feat: support gutentags_file_list_command on the gtags_cscope module

The original gutentags_file_list_command setting only works on ctags and cscope module, so to support the gtags_cscope module can be achieved through its -f option.

---

* feat: provide the GutentagsRoot command

Once Gutentags completed the locating of the project root, the user can change the directory to it via the GutentagsRoot command.

---

* feat: provide the g:gutentags_root_located_user_func hook

Once Gutentags complete the locating of the project root, it will call this user function (if present) to allow the user to load some custom configurations in the project root.

---

* feat: change the behavior of the g:gutentags_generate_on_new setting

In the original version, if the tag file does not exist and the g:gutentags_generate_on_missing is false, the g:gutentags_generate_on_new will control whether or not to generate a new tag. In general, this behavior is unpredictable for user who settings g:gutentags_generate_on_missing to False.

Therefore, let g:gutentags_generate_on_new setting only work when the tag file exists.